[CXF-7022] Feature implemented
